Move sync folder
